litex_mmc: ensure bus_width is (re)configured upon card (re)insertion
Before get_cd() is about to return 0 (card not present), reset the
`host->is_bus_width_set` to `false`, so that when we subsequently
re-insert the card, `bus_width` will be configured again prior to
attempting to use the card. This allows the driver to correctly
handle card eject and re-insert, beyond the very first iteration.

Signed-off-by: Gabriel Somlo <gsomlo@gmail.com>