litex_mmc: clarify how/why bus-width is configured
LiteSDCard does not support the "standard" 1-bit bus width that cards
(and the Linux mmc subsystem) both use as a default. As such, the card's
bus-width must be set to 4-bit (the only width supported by LiteSDCard)
before the very first data transfer occurs (which is typically much earlier
than when the Linux mmc subsystem would switch to "wide mode".

Also, if we're only supporting one bus width, there's no point in storing
it as a variable in the "litex_mmc_host" data structure.

Signed-off-by: Gabriel Somlo <gsomlo@gmail.com>