litex-mmc: FIXME: additional halving of the sdclock frequency
When the reference clock is an exact power-of-two multiple of the
requested sdclock, we can "honor" the fastest requested value. This
sometimes ends up being "too fast", as observed by the inability to
boot from sdcard in the bios.

Until we get to the bottom of why that happens, or determine it's not
a concern in Linux, let's additionally halve the sdclock to ensure it's
"slow enough" not to "strain" the gateware or physical card on the other
end...

NOTE: doing this will not cause a perceptible increase in the reported
time for dd-ing a large (15MB) file from a mounted sdcard partition to
/dev/null (22 seconds on LiteX+Rocket).

Signed-off-by: Gabriel Somlo <gsomlo@gmail.com>