dt-bindings: gpio: litex,gpio: Convert to json-schema
Convert the LiteX GPIO controller Device Tree binding documentation to
json-schema.

  - Add SPDX license,
  - Document missing properties,
  - Document GPIO hogs,
  - Fix reg properties in examples, which are always built with
    #{address,size}-cells = <1>.

Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>
---
The original file didn't contain license information, so I used the most
popular/recommended one for DT bindings.
I listed the original author of the bindings as maintainer.

Interrupt support is not included, as this is not yet supported by the
Linux driver.