litex_mmc: Tell the controller when we need to wait while the card is busy
This sets the response type in the CMDCMD register to
SDCARD_CTRL_RESPONSE_SHORT_BUSY (i.e. 3) when the upper layers
indicate that the current command could cause the card to signal that
is busy by pulling the DAT0 line low, and that we need to wait for
DAT0 to go high before indicating that the command is complete.

With this, erase commands work successfully.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>