litex_mmc: Use interrupts to signal command completion
For commands that transfer data or where the card can signal a busy
status (via DAT0), this uses the command-completion interrupt together
with a struct completion to wait for the command to finish rather than
spinning.  This allows other tasks to be scheduled while the card is
executing long-running commands.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>