irqchip: LITEX_VEXRISCV_INTC should depend on LITEX && RISCV
The LiteX VexRiscv interrupt controller is only present on VexRiscv
SoCs, and its driver uses csr_*() accessor functions that are available
on RISCV only.  Hence add dependencies on LITEX and RISCV, to prevent
asking the user about this driver when configuring a kernel without
LiteX or RISC-V support.

Fixes: 81635b3a759af145 ("LiteX: support for VexRiscV interrupt controller")
Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>
---
Feel free to fold into the original commit.